Code for Dendrograms and VertexClustering
Fix issue #86
