+++ data/sakai-devel/community_vertex_clustering.txt
@@ -1,0 +1,36 @@
+Clustering with 100 elements and 2 clusters
+[0] ggolden@umich.edu, sakai-dev@collab.sakaiproject.org, csev@umich.edu,
+    john@caret.cam.ac.uk, ys2n@virginia.edu, s-githens@northwestern.edu,
+    slt@columbia.edu, ajpoland@iupui.edu, aaronz@vt.edu, caseyd1@stanford.edu,
+    sakai-user@collab.sakaiproject.org, john.ellis@rsmart.com,
+    dave.ross@gmail.com, clayf@bu.edu, ray@media.berkeley.edu,
+    jleasia@umich.edu, antranig@caret.cam.ac.uk, bkirschn@umich.edu,
+    vrajgopalan@ucmerced.edu, jholtzman@berkeley.edu, zqian@umich.edu,
+    marquard@ched.uct.ac.za, nuno@ufp.pt, jimeng@umich.edu,
+    markjnorton@earthlink.net, arwhyte@umich.edu, kimkyong@fhda.edu,
+    tpamsler@ucdavis.edu, dhorwitz@ched.uct.ac.za, lance@indiana.edu,
+    lydial@stanford.edu, knoop@umich.edu, hedrick@rutgers.edu,
+    dlhaines@umich.edu, mmmay@indiana.edu, sakai-qa@collab.sakaiproject.org,
+    brian@moosetrout.com, jpgorrono@ucdavis.edu, chmaurer@iupui.edu,
+    omer@rice.edu, ian@caret.cam.ac.uk, njbotime@svsu.edu, ieb@tfd.co.uk,
+    ottenhoff@longsight.com, production@collab.sakaiproject.org,
+    mustansar@rice.edu, SinouVivie@foothill.edu, mbreuker@loi.nl,
+    rpaditya@umich.edu, A.M.Berg@uva.nl, J.Higham@hull.ac.uk,
+    will@serensoft.com, harriet@caret.cam.ac.uk, da1@vt.edu, smkeesle@syr.edu,
+    ktsao@stanford.edu, matthew.buckett@oucs.ox.ac.uk, a.fish@lancaster.ac.uk,
+    david.horwitz@uct.ac.za, stephen.marquard@uct.ac.za, cdunstall@csu.edu.au,
+    darolmar@upvnet.upv.es, kchan@hawaii.edu, sswinsb2@une.edu.au,
+    sgithens@caret.cam.ac.uk, michael.feldstein@oracle.com,
+    jwheat@messiah.edu, adam.marshall@oucs.ox.ac.uk, john.bush@rsmart.com,
+    clay.fenlason@et.gatech.edu, mkorcuska@gmail.com, botimer@umich.edu,
+    sakai-ux@collab.sakaiproject.org, me@nathanpearson.com,
+    daniel.merino@unavarra.es, kpfoote@iup.edu, alaanizar@iihem.ac.ma, jean-
+    francois.leveque@upmc.fr, nuno@ufp.edu.pt, jonespm@umich.edu,
+    swgithen@mtu.edu, s.swinsburg@lancaster.ac.uk, joshua.swink@gmail.com,
+    farreri.sakai@gmail.com, steve.swinsburg@gmail.com,
+    warwickchapman@gmail.com, azeckoski@unicon.net,
+    cmw6s@eservices.virginia.edu, holladay@longsight.com, davidwaf@gmail.com,
+    matthew@longsight.com, euksa99@gmail.com, adrian.r.fish@gmail.com,
+    nealcaidin@sakaifoundation.org, adam.marshall@it.ox.ac.uk,
+    neal.caidin@apereo.org, sakai-dev@apereo.org
+[1] LvDLADg@public.gmane.org, A@public.gmane.org, yROfE0A@public.gmane.org
